backend/requirements.txt
fastapi
uvicorn
pandas
pyarrow
sqlalchemy
python-multipart
python-dotenv
